Move in the new aggregators
